No longer have a `routeToUri` function in the router.
Instead, a router that implements `uriToRoute()` can choose to add it's own route-to-uri helpers.

Why?

The current signiature allows for passing in routes that don't exist, and so the return type is `Option<String>` rather than `String`.

I found in practice this is quite annoying to use, and having a different approach allows us to sidestep it.

So instead of

    router.routeToUri(SearchPage, {query: "ice-cream"}

We could have a helper like:

    router.uriForSearchPage({query: "ice-cream"});

This way the type system avoids ever allowing impossible routes to be requested.

And because the route->uri pathway is only really needed by consuming code (the framework itself never needs it) removing the function is a non issue.

Again, I'm hoping that we can auto-generate these routing classes with a fancy macro.