Add Snabbdom, and add a weird way of making it available
I really couldn't be bothered with webpack (or parcel or es-build etc). Snabbdom is ready for modern browsers: it's plain JavaScript, but using ES Modules.

So as long as we use `<script type="module>` then we can `import {} from "snabbdom/..."`.

I've added a "jsImport" haxe library to produce the ES6 import syntax, and made the externs load from a specific URL that I created on our server to share the files.

It's pretty weird, but I'm grateful not to be including something like webpack this early in the project. I can see myself reversing this and adding a bundle step later.