Introduce esbuild as a bundler
And get rid of the esmodules-from-the-server thing.

ESBuild is simple enough and fast enough (compared to webpack etc) that
I'm happy.

And the previous trick was clever, but fragile. It didn't work in node (server
side or unit tests) and the browser with the same lines of code. And it wouldn't
work for other NPM packages as easily.