Add an initial Meals domain model + EventSource
Note: I've run into some pretty horrible needs for conditional compilation that I'd like to resolve next, because the MealsListApi (for example) needs to reference a MealsEventSource which can only be compiled on the server because of various NodeJS imports. But because our routing has a `Page()` object that requires to know the API, and the routing is used client side... it gets real messy.

I'm going to try clean this up next.