Refactor to `Page` types to allow `PageApi` to be kept out of the client
This allows us to separate PageApi files from anything that is loaded on the client, so they can safely use all sorts of server specific code and not worry about client side compilation errors.

See the diff in `domains/Meals.hx` for an example of the annoying workarounds that are no longer required.

This refactor also makes the `handlePage()` function in NodeJS.hx easier because the stuff needed for a page is more self-contained in the Page object.

It also leaves the door open for other non-page routes (eg APIs, downloads, RSS, iCal etc) in future.