Add support for render lifecycle hooks
These are based on those supplied by Snabbdom, though I've wrapped the API and only exposed a few for now.

Interested to see which things feel the most useful.

I'm also interested to see if we can make them easy to compose (creating the equivalent of React's `useEffect` for example)