Move responsibility for querying data after a command out of `doCommand`
Now its in handleRequestForPage(). I think the fact `doCommand()` also did the query, and "CQRS" is a term, was bugging me.

The reality that this implementation is still waiting synchronously on the command results before doing a query and returning PageData in the same HttpRequest has not changed, but I feel like its in a less offensive position now.