Update `Init` hook to allow returning a cleanup function
Changes:
- Make `Init` hook callback expect to return `Option<Callback>`
- Remove `onInitAndDestroy`
- Add internal/private APIs for persisting some state on particular nodes between renders. We aren't exposing any local/private state controls through our APIs yet, but we're using this to implement the `Init` cleanup callbacks.

The previous `onInitAndDestroy` function did not actually work, and was causing errors in our mealplanner example app when navigating away from the ShoppingList page, which should have triggered a `clearInterval` but did not.

The `onInitAndDestroy` function is actually called every time the virtual dom is generated and so the `var cleanup` is captured separately once per render call, and does not persist between renders ... kind of like the classic React setInterval problem.

In this library they use `data` on the VNode, and keep transferring it to the new VNode in an `update` hook:
https://github.com/alfonsogarciacaro/Feliz.Snabbdom/blob/main/src/Feliz.Snabbdom/Feliz.Snabbdom.fs#L141-L157

We've done something similar - setting up a data structure that can be persisted between vNodes in separate renders, and using it to store the cleanup functions from our `Init` hooks, and then call each of the cleanup functions during the `Destroy` phase for the node.